changed to channel in api and altered tags
