--- src/main/java/uk/gov/hmcts/reform/channel/management/errorhandling/ExceptionResponse.java
@@ -1,23 +1,0 @@
-package uk.gov.hmcts.reform.channel.management.errorhandling;
-
-import lombok.Data;
-
-import java.time.LocalDateTime;
-
-/**
- * Exception Response class, to standardise exceptions returned from the server.
- */
-@Data
-public class ExceptionResponse {
-
-    /**
-     * The error message to return.
-     */
-    private String message;
-
-    /**
-     * The timestamp of when the error occurred.
-     */
-    private LocalDateTime timestamp;
-
-}
