--- src/main/java/uk/gov/hmcts/reform/channel/management/errorhandling/GlobalExceptionHandler.java
@@ -1,37 +1,0 @@
-package uk.gov.hmcts.reform.channel.management.errorhandling;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.context.request.WebRequest;
-import uk.gov.hmcts.reform.channel.management.errorhandling.exceptions.ChannelNotFoundException;
-
-import java.time.LocalDateTime;
-
-/**
- * Global exception handler, that captures exceptions thrown by the controllers, and encapsulates
- * the logic to handle them and return a standardised response to the user.
- */
-@ControllerAdvice
-public class GlobalExceptionHandler {
-
-    /**
-     * Template exception handler, that handles a custom ChannelNotFoundException,
-     * and returns a 404 in the standard format.
-     * @param ex The exception that has been thrown.
-     * @param request The request made to the endpoint.
-     * @return The error response, modelled using the ExceptionResponse object.
-     */
-    @ExceptionHandler(ChannelNotFoundException.class)
-    public ResponseEntity<ExceptionResponse> handleChannelNotFound(
-        ChannelNotFoundException ex, WebRequest request) {
-
-        ExceptionResponse exceptionResponse = new ExceptionResponse();
-        exceptionResponse.setMessage(ex.getMessage());
-        exceptionResponse.setTimestamp(LocalDateTime.now());
-
-        return ResponseEntity.status(HttpStatus.NOT_FOUND).body(exceptionResponse);
-    }
-
-}
