--- src/main/java/uk/gov/hmcts/reform/channel/management/models/external/subscriptionmanagement/Channel.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.channel.management.models.external.subscriptionmanagement;
-
-public enum Channel {
-    // add as required
-    EMAIL,
-    API;
-}
