+++ src/main/java/uk/gov/hmcts/reform/pip/channel/management/Application.java
@@ -1,0 +1,13 @@
+package uk.gov.hmcts.reform.pip.channel.management;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+@SuppressWarnings("HideUtilityClassConstructor") // Spring needs a constructor, its not a utility class
+public class Application {
+
+    public static void main(final String[] args) {
+        SpringApplication.run(Application.class, args);
+    }
+}
