+++ src/main/java/uk/gov/hmcts/reform/pip/channel/management/models/external/subscriptionmanagement/Channel.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.pip.channel.management.models.external.subscriptionmanagement;
+
+public enum Channel {
+    // add as required
+    EMAIL,
+    API;
+}
