--- src/test/java/uk/gov/hmcts/reform/channel/management/errorhandling/exceptions/ChannelNotFoundExceptionTest.java
@@ -1,19 +1,0 @@
-package uk.gov.hmcts.reform.channel.management.errorhandling.exceptions;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-class ChannelNotFoundExceptionTest {
-
-    @Test
-    void testCreationOfSubscriptionNotFoundException() {
-
-        ChannelNotFoundException subscriptionNotFoundException
-            = new ChannelNotFoundException("This is a test message");
-        assertEquals("This is a test message", subscriptionNotFoundException.getMessage(),
-                     "The message should match the message passed in");
-
-    }
-
-}
